AMEN
The Timer was extended to provide an event on every 
second passed.

The UI Controller now uses the timer for the round countdown HUD.

The HUD was slightly rewritten to update the text only when a variable was changed.

There are issues when trying to eliminate players. That code has been commented out.